junel/requirements.txt
opencv-python==4.8.1.78
tensorflow==2.13.0
numpy==1.24.3
imutils==0.5.4
paho-mqtt==1.6.1
psycopg2-binary==2.9.7
requests==2.31.0
python-telegram-bot==20.4
redis==4.6.0
celery==5.3.1
flask==2.3.2
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
schedule==1.2.0
pillow==10.0.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
pandas==2.0.3
